Refactor indexer tests to avoid duplicate code
