+++ statediff/indexer/database/file/mainnet_tests/indexer_test.go
@@ -30,6 +30,7 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/ethereum/go-ethereum/statediff/indexer/database/file"
+	"github.com/ethereum/go-ethereum/statediff/indexer/database/sql"
 	"github.com/ethereum/go-ethereum/statediff/indexer/database/sql/postgres"
 	"github.com/ethereum/go-ethereum/statediff/indexer/interfaces"
 	"github.com/ethereum/go-ethereum/statediff/indexer/mocks"
@@ -37,6 +38,7 @@ import (
 )
 
 var (
+	db        sql.Database
 	sqlxdb    *sqlx.DB
 	chainConf = params.MainnetChainConfig
 )
@@ -126,7 +128,7 @@ func dumpData(t *testing.T) {
 }
 
 func tearDown(t *testing.T) {
-	file.TearDownDB(t, sqlxdb)
+	test_helpers.TearDownDB(t, db)
 	err := os.Remove(file.CSVTestConfig.FilePath)
 	require.NoError(t, err)
 	err = sqlxdb.Close()
